FastAPI Tutorial/pyproject.toml
[project]
name = "fastapi-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "boto3>=1.42.89",
    "fastapi>=0.135.3",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "imagekitio>=5.3.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.44.0",
]
